Export creation of collections from files
