+++ index.js
@@ -1,4 +1,4 @@
-const { ActionRowBuilder, Client, Collection, GatewayIntentBits, Partials } = require("discord.js");
+const { Client, GatewayIntentBits, Partials } = require("discord.js");
 const { token } = require("./config.json");
 const path = require("path");
 const fs = require("fs");
@@ -8,59 +8,17 @@ const client = new Client({
 	partials: [Partials.Channel, Partials.Message, Partials.Reaction],
 });
 
-const foldersName = ["commands", "buttons", "modals"];
+// Create Collections for Buttons, Commands and Modals.
+const { createInteractionCollections } = require("./utils/createInteractionCollections.js");
+createInteractionCollections(client);
 
-for (const folderName of foldersName) {
-	client[folderName] = new Collection();
-	const folderPath = path.join(__dirname, folderName);
-	const files = fs.readdirSync(folderPath).filter((file) => file.endsWith(".js"));
+// Create Message Action (Buttons) for messageCreate interaction
+const { createMessageAction } = require("./utils/createMessageAction.js");
+createMessageAction(client);
 
-	for (const file of files) {
-		const filePath = path.join(folderPath, file);
-		const content = require(filePath);
-
-		if (folderName === "commands") {
-			client[folderName].set(content.data.name, content);
-		} else {
-			client[folderName].set(content.data.toJSON()?.custom_id, content);
-		}
-	}
-}
-
-// Sort buttons by indice
-const msgActionRow = new ActionRowBuilder();
-// @ts-ignore
-[...client.buttons.values()]
-	.sort((A, B) => A.indice - B.indice)
-	.forEach((button) => msgActionRow.addComponents(button.data));
-
-// @ts-ignore
-client.messageAction = new Collection();
-// @ts-ignore
-client.messageAction.set("messageActionButton", msgActionRow);
-
-// Reading reactions from file
-// @ts-ignore
-client.reactions = new Collection();
-const reactionsPath = path.join(__dirname, "./data/reactions.json");
-const reactionsFile = JSON.parse(fs.readFileSync(reactionsPath, { encoding: "utf-8" }));
-
-// @ts-ignore
-Object.keys(reactionsFile).forEach((msgId) => client.reactions.set(msgId, reactionsFile[msgId]));
-
-// Reading authorized channels from file
-// @ts-ignore
-client.authorizedChannels = new Collection();
-const channelsPath = path.join(__dirname, "./data/channels.json");
-// @ts-ignore
-client.authorizedChannels = JSON.parse(fs.readFileSync(channelsPath, { encoding: "utf-8" }));
-
-// Reading logs configuration parameters from file
-// @ts-ignore
-client.logsConfiguration = new Collection();
-const logsConfigurationPath = path.join(__dirname, "./data/logsConfiguration.json");
-// @ts-ignore
-client.logsConfiguration = JSON.parse(fs.readFileSync(logsConfigurationPath, { encoding: "utf-8" }));
+// Create Collections for data folder.
+const { createDataCollections } = require("./utils/createDataCollections.js");
+createDataCollections(client);
 
 // Events
 const eventsPath = path.join(__dirname, "events");
